+++ tools/ship-services.sh
@@ -2,12 +2,12 @@
 
 set -e
 
-provisioning/start-staging.sh
+deploy/start-staging.sh
 
 tools/clean-workspace.sh
 tools/build-it.sh --console=plain --no-daemon --no-build-cache --max-workers=1 build
 
-provisioning/deploy-login.sh
-provisioning/deploy-dashboard.sh
+deploy/deploy-login.sh
+deploy/deploy-dashboard.sh
 
-provisioning/stop-staging.sh
+deploy/stop-staging.sh